Added the UI for Icon and IconSet details as a Fragment. This commit has some TODOs and comments which are to deleted during the final implementation of these screens. Added styles to the download button to match it with the given UI.
