+++ app/src/main/java/com/srilasaka/iconfinderapp/ui/icon_details/IconDetailsFragment.kt
@@ -1,0 +1,40 @@
+package com.srilasaka.iconfinderapp.ui.iconset_details
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.srilasaka.iconfinderapp.databinding.FragmentIconDetailsBinding
+
+/**
+ * [Fragment] to display the details od the IconSet from [IconsFragment].
+ */
+class IconDetailsFragment : Fragment() {
+
+    private var _binding: FragmentIconDetailsBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+        _binding = FragmentIconDetailsBinding.inflate(inflater, container, false)
+        return binding.root
+
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
