+++ app/src/main/java/com/srilasaka/iconfinderapp/ui/iconset_details/IconSetDetailsFragment.kt
@@ -1,0 +1,40 @@
+package com.srilasaka.iconfinderapp.ui.iconset_details
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.srilasaka.iconfinderapp.databinding.FragmentIconSetDetailsBinding
+
+/**
+ * [Fragment] to display the details od the IconSet from [IconSetFragment].
+ */
+class IconSetDetailsFragment : Fragment() {
+
+    private var _binding: FragmentIconSetDetailsBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+
+        _binding = FragmentIconSetDetailsBinding.inflate(inflater, container, false)
+        return binding.root
+
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
