+++ app/src/main/res/layout/recycler_item_icon_set.xml
@@ -131,7 +131,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:orientation="vertical"
-                    app:layout_constraintGuide_percent="0.45"/>
+                    app:layout_constraintGuide_begin="171dp" />
 
                 <TextView
                     android:id="@+id/tv_icon_set_license_type_title"
