+++ app/src/main/res/values/colors.xml
@@ -13,10 +13,12 @@
 
     <color name="button_color">@color/purple</color>
     <color name="text_color">#5E6988</color>
+    <color name="text_color_purple">@color/purple</color>
     <color name="search_bg_color">#F3F5F6</color>
     <color name="card_bg_color">#F1F2F5</color>
     <color name="image_bg_color">#00000012</color>
     <color name="toolbarBackgroundColor">@color/white</color>
     <color name="ripple_color_grey">#593B3B3B</color>
+    <color name="ripple_color_purple">@color/purple_500</color>
 
 </resources>
