+++ app/src/main/res/values/strings.xml
@@ -37,4 +37,7 @@
     <string name="select">Select</string>
     <string name="premium">Premium</string>
     <string name="free">Free</string>
+    <string name="sample_string_website_url">www.icon7.com</string>
+    <string name="info">Info:</string>
+    <string name="sample_string_info">"At vero eos et accusamus et iusto odi dignissimos ducimus qui blanditiis praesentium voluptatum deleniti atq ue corrupti quos dolores et quas mol estias excepturi sint occaecati cupidi. praesentium voluptatum deleniti atq ue corrupti quos dolores et quas mol. "</string>
 </resources>
